Commit the fixes for real! and renamed the files to just _test.py so that pytest can find them.
