+++ tests/runner/local_runner_test.py
@@ -2,7 +2,7 @@ from blurr.runner.local_runner import LocalRunner
 
 
 def test_local_runner():
-    local_runner = LocalRunner(['data/raw.json'], '', 'data/sample.yml')
+    local_runner = LocalRunner(['tests/runner/data/raw.json'], '', 'tests/runner/data/sample.yml')
     local_runner.execute()
 
     assert local_runner._user_transformer['userA'].snapshot['session'][
